refactor: create method and delete not necessary function
